Extracted the new interfaces, fixed an old one, used delegation of old to (intended) new classes - all names are prelimenary.
